Made the SideLength that is used configurable so that you can change sizes of the objects drawn.

+++ OOPDraw/Heptagon.cs
@@ -4,7 +4,7 @@ namespace OOPDraw
 {
     internal class Heptagon : RegularPolygon
     {
-        public Heptagon(int xOrigin, int yOrigin, Color colour, float penSize, int sideLength) : base(xOrigin, yOrigin, colour, penSize, sideLength, 7)
+        public Heptagon(int xOrigin, int yOrigin, Color colour, float penSize, float sideLength) : base(xOrigin, yOrigin, colour, penSize, sideLength, 7)
         {
         }
     }
