+++ OOPDraw/Hexagon.cs
@@ -4,7 +4,7 @@ namespace OOPDraw
 {
     internal class Hexagon : RegularPolygon
     {
-        public Hexagon(int xOrigin, int yOrigin, Color colour, float penSize, int sideLength) : base(xOrigin, yOrigin, colour, penSize, sideLength, 6)
+        public Hexagon(int xOrigin, int yOrigin, Color colour, float penSize, float sideLength) : base(xOrigin, yOrigin, colour, penSize, sideLength, 6)
         {
         }
     }
